move unstack to manipulation, one_hot to creation
