+++ ivy/container/general.py
@@ -1313,164 +1313,6 @@ class ContainerWithGeneral(ContainerBase):
         )
 
     @staticmethod
-    def static_unstack(
-        x: Union[ivy.Array, ivy.NativeArray, ivy.Container],
-        axis: int,
-        /,
-        *,
-        keepdims: bool = False,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-    ) -> ivy.Container:
-        """
-        ivy.Container static method variant of ivy.unstack. This method
-        simply wraps the function, and so the docstring for ivy.unstack
-        also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        x
-            Input array or container to unstack.
-        axis
-            Axis for which to unpack the array.
-        keepdims
-            Whether to keep dimension 1 in the unstack dimensions. Default is False.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is None.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains
-            will be skipped. Default is True.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied.
-            Default is False.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples). Default is False.
-
-        Returns
-        -------
-        ret
-            List of arrays, unpacked along specified dimensions, or containers
-            with arrays unpacked at leaves
-
-        Examples
-        --------
-        With one :code:`ivy.Container` input:
-
-        >>> x = ivy.Container(a=ivy.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]]]),
-                            b=ivy.array([[[9, 10], [11, 12]], [[13, 14], [15, 16]]]))
-        >>> y = ivy.Container.static_unstack(x, axis=0)
-        >>> print(y)
-        [{
-            a: ivy.array([[1, 2],
-                         [3, 4]]),
-            b: ivy.array([[9, 10],
-                         [11, 12]])
-        }, {
-            a: ivy.array([[5, 6],
-                         [7, 8]]),
-             b: ivy.array([[13, 14],
-                          [15, 16]])
-        }]
-
-        >>> x = ivy.Container(a=ivy.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]]]),
-                            b=ivy.array([[[9, 10], [11, 12]], [[13, 14], [15, 16]]]))
-        >>> y = ivy.Container.static_unstack(x, axis=1, keepdims=True)
-        >>> print(y)
-        [{
-            a: ivy.array([[[1, 2]],
-                         [[5, 6]]]),
-            b: ivy.array([[[9, 10]],
-                         [[13, 14]]])
-        }, {
-            a: ivy.array([[[3, 4]],
-                         [[7, 8]]]),
-            b: ivy.array([[[11, 12]],
-                         [[15, 16]]])
-        }]
-        """
-        return ContainerBase.multi_map_in_static_method(
-            "unstack",
-            x,
-            axis,
-            keepdims=keepdims,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-        )
-
-    def unstack(
-        self: ivy.Container,
-        axis: int,
-        /,
-        *,
-        keepdims: bool = False,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-    ) -> ivy.Container:
-        """
-        ivy.Container instance method variant of ivy.unstack. This method
-        simply wraps the function, and so the docstring for ivy.unstack
-        also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        self
-            Input container to unstack at leaves.
-        axis
-            Axis for which to unpack the array.
-        keepdims
-            Whether to keep dimension 1 in the unstack dimensions. Default is False.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is None.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains
-            will be skipped. Default is True.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied.
-            Default is False.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples). Default is False.
-
-        Returns
-        -------
-        ret
-            Containers with arrays unpacked at leaves
-
-        Examples
-        --------
-        With one :code:`ivy.Container` instances:
-
-        >>> x = ivy.Container(a=ivy.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]]]),
-                            b=ivy.array([[[9, 10], [11, 12]], [[13, 14], [15, 16]]]))
-        >>> x.unstack(axis=0)
-        [{
-            a: ivy.array([[1, 2],
-                         [3, 4]]),
-            b: ivy.array([[9, 10],
-                          [11, 12]])
-        }, {
-            a: ivy.array([[5, 6],
-                          [7, 8]]),
-            b: ivy.array([[13, 14],
-                          [15, 16]])
-        }]
-        """
-        return self.static_unstack(
-            self,
-            axis,
-            keepdims=keepdims,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-        )
-
-    @staticmethod
     def static_fourier_encode(
         x: Union[ivy.Container, ivy.Array, ivy.NativeArray],
         max_freq: Union[float, ivy.Array, ivy.NativeArray],
@@ -3420,106 +3262,6 @@ class ContainerWithGeneral(ContainerBase):
         """
         return self.static_indices_where(
             self,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-            out=out,
-        )
-
-    @staticmethod
-    def static_one_hot(
-        indices: ivy.Container,
-        depth: int,
-        /,
-        *,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-        out: Optional[ivy.Container] = None,
-    ) -> ivy.Container:
-        """
-        ivy.Container static method variant of ivy.one_hot. This method
-        simply wraps the function, and so the docstring for ivy.one_hot
-        also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        indices
-            Indices for where the ones should be scattered *[batch_shape, dim]*
-        depth
-            Scalar defining the depth of the one-hot dimension.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is None.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains will
-            be skipped. Default is True.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied. Default
-            is False.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples). Default is False.
-
-        Returns
-        -------
-        ret
-            container with tensors of zeros with the same shape and type as the inputs,
-            unless dtype provided which overrides.
-        """
-        return ContainerBase.multi_map_in_static_method(
-            "one_hot",
-            indices,
-            depth,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-            out=out,
-        )
-
-    def one_hot(
-        self: ivy.Container,
-        depth: int,
-        /,
-        *,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-        out: Optional[ivy.Container] = None,
-    ):
-        """
-        ivy.Container instance method variant of ivy.one_hot. This method
-        simply wraps the function, and so the docstring for ivy.one_hot
-        also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        self
-            Indices for where the ones should be scattered *[batch_shape, dim]*
-        depth
-            Scalar defining the depth of the one-hot dimension.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is None.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains will
-            be skipped. Default is True.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied. Default
-            is False.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples). Default is False.
-
-        Returns
-        -------
-        ret
-            container with tensors of zeros with the same shape and type as the inputs,
-            unless dtype provided which overrides.
-        """
-        return self.static_one_hot(
-            self,
-            depth,
             key_chains=key_chains,
             to_apply=to_apply,
             prune_unapplied=prune_unapplied,
