+++ ivy/functional/backends/numpy/manipulation.py
@@ -202,6 +202,15 @@ def swapaxes(
     return np.swapaxes(x, axis0, axis1)
 
 
+def unstack(x: np.ndarray, axis: int, keepdims: bool = False) -> List[np.ndarray]:
+    if x.shape == ():
+        return [x]
+    x_split = np.split(x, x.shape[axis], axis)
+    if keepdims:
+        return x_split
+    return [np.squeeze(item, axis) for item in x_split]
+
+
 def clip(
     x: np.ndarray,
     x_min: Union[Number, np.ndarray],
