+++ ivy/functional/backends/torch/manipulation.py
@@ -280,3 +280,12 @@ def clip(
 
 clip.support_native_out = True
 clip.unsupported_dtypes = ("float16",)
+
+
+def unstack(x: torch.Tensor, axis: int, keepdims: bool = False) -> List[torch.Tensor]:
+    if x.shape == ():
+        return [x]
+    ret = list(torch.unbind(x, axis))
+    if keepdims:
+        return [r.unsqueeze(axis) for r in ret]
+    return ret
